refactor: reorganize complex example into packages
