+++ 03-Karmasik/gui/__init__.py
@@ -1,0 +1,1 @@
+from .base_gui import EnterpriseGUI
