+++ 03-Karmasik/gui/aesthetic_engine.py
@@ -1,0 +1,5 @@
+"""Aesthetic engine functions for GUI styling."""
+
+class AestheticEngine:
+    """Placeholder mixin for aesthetic-related methods."""
+    pass
