+++ 03-Karmasik/gui/animations.py
@@ -1,0 +1,5 @@
+"""Animation helpers for GUI."""
+
+class AnimationEngine:
+    """Placeholder mixin for animations."""
+    pass
