+++ 03-Karmasik/gui/modal_system.py
@@ -1,0 +1,5 @@
+"""Modal windows for data entry and dialogs."""
+
+class ModalSystem:
+    """Placeholder mixin for modal dialogs."""
+    pass
