+++ 03-Karmasik/gui/styles.py
@@ -1,0 +1,6 @@
+"""Style constants for the GUI."""
+
+DEFAULT_STYLE = {
+    "bg": "#0d1117",
+    "fg": "#cdd6f4",
+}
