+++ 03-Karmasik/rpa/excel_processor.py
@@ -1,0 +1,10 @@
+"""Utilities for reading and processing Excel files."""
+
+import pandas as pd
+from pathlib import Path
+from typing import List
+
+
+def read_excel_files(files: List[Path]):
+    """Placeholder Excel processing."""
+    return [pd.read_excel(f) for f in files]
